Move the phone's composer and start screen out of main
Two more clusters leave main.rs. composer.rs holds what sits under the prompt
box — the model and effort pickers, and what submitting the prompt does.
start.rs holds the screen the phone opens on, the pair card it falls back to
when there is no Mac yet, and the two launcher rows it draws.

What is left in main.rs is the launch glue and AppBody, which is still 800
lines and still the real problem: one component branching on twenty-five
signals between four screens. Extracting the chat screen from it means giving
those signals an owner, which is a refactor rather than a move.
